+++ src/components/common/Buttons/EditMessageButton.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useNavigate, useParams } from 'react-router-dom';
-import { Primary56 } from './ButtonStyle';
+import { Primary56 } from '../../../styles/ButtonStyle';
 
 function EditMessageButton({ messageID }) {
   const navigate = useNavigate();
