+++ src/components/common/Buttons/movePageButton.jsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import { useNavigate } from 'react-router-dom';
-import { Primary56 } from './ButtonStyle';
+import { Primary56 } from '../../../styles/ButtonStyle';
 
 function MovePageButton({ moveLink, btnName }) {
   const navigate = useNavigate();
